+++ tests/kltests.cc
@@ -2,7 +2,7 @@
 #include "kltests.h"
 
 const char *SAMPLE_MESSAGE = "Hello World 1234567890";
-const size_t SAMPLE_MESSAGE_LEN = strlen(SAMPLE_MESSAGE);
+const UInt64 SAMPLE_MESSAGE_LEN = strlen(SAMPLE_MESSAGE);
 
 void makeRandomMessage(char *buffer, int msgId, int nRand)
 {
@@ -18,9 +18,9 @@ void makeRandomMessage(char *buffer, int msgId, int nRand)
 long long publishMessages(KLTopic *topic, int numMessages, int nRandMax, bool assertOffsets)
 {
 	// long long beforeTime = kl_current_timestamp();
-	size_t offset = 0;
-	size_t totalSize = 0;
-	size_t msgsize = SAMPLE_MESSAGE_LEN;
+	UInt64 offset = 0;
+	UInt64 totalSize = 0;
+	UInt64 msgsize = SAMPLE_MESSAGE_LEN;
 	const char *message = SAMPLE_MESSAGE;
 	char buffer[4096];
 	long long elapsedTime = 0;
@@ -40,10 +40,10 @@ long long publishMessages(KLTopic *topic, int numMessages, int nRandMax, bool as
 		totalSize += msgsize;
 		if (assertOffsets)
 		{
-			CPPUNIT_ASSERT(offset == (totalSize + (sizeof(size_t) * (i + 1))));
+			CPPUNIT_ASSERT(offset == (totalSize + (sizeof(UInt64) * (i + 1))));
 		}
 	}
 	CPPUNIT_ASSERT(kl_topic_message_count(topic) == numMessages);
-	CPPUNIT_ASSERT(totalSize + (sizeof(size_t) * numMessages) == offset);
+	CPPUNIT_ASSERT(totalSize + (sizeof(UInt64) * numMessages) == offset);
 	return elapsedTime;
 }
